libpayload: Fix building dwc2 UDC driver with debug enabled
hexdump() now takes a pointer instead of an int-containing-an-address.

BUG=none
BRANCH=none
TEST=building with USB_DEBUG works

Original-Commit-Id: 8660f6091bb124eeabe73302e8c7f1a8e46324f1
Original-Change-Id: I266efcb8b939d6da104ad05a3e79a78065c60beb
Original-Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/308762
Original-Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Original-Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Original-Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Original-Reviewed-by: yunzhi li <lyz@rock-chips.com>
Reviewed-on: http://review.coreboot.org/12258
Tested-by: build bot (Jenkins)

+++ payloads/libpayload/drivers/udc/dwc2.c
@@ -563,7 +563,7 @@ static void dwc2_outep_intr(struct usbdev_ctrl *this, dwc2_ep_t *ep)
 		usb_debug("DEPINT_SETUP\n");
 		writel(DXEPINT_SETUP, &ep->ep_regs->depint);
 #ifdef USB_DEBUG
-		hexdump((unsigned int)p->setup_buf, sizeof(dev_req_t));
+		hexdump(p->setup_buf, sizeof(dev_req_t));
 #endif
 		SIMPLEQ_REMOVE_HEAD(&p->eps[0][0].job_queue, queue);
 		p->eps[0][0].busy = 0;
